training/train.py: modify train loop + eval to work with new VideoCLIP wrapper

+++ src/aggregation/aggregator_wrapper.py
@@ -13,7 +13,7 @@ class VideoCLIP(nn.Module):
         self.logit_scale = torch.nn.Parameter(torch.ones([]) * np.log(1 / 0.07))
 
     def forward(self, x, prenorm=True, postnorm=True):
-        x = F.normalize(x, dim=-1) if prenorm else x
+        x = F.normalize(x.float(), dim=-1) if prenorm else x
         x = self.aggregator(x)
-        x = F.normalize(x, dim=-1) if postnorm else x
+        x = F.normalize(x.float(), dim=-1) if postnorm else x
         return x, self.logit_scale.exp()
